homework5/streamlit/requirements.txt
streamlit
pandas
psycopg2-binary
plotly
requests
python-dotenv
bcrypt
extra-streamlit-components
pytz
